+++ src/winimg.c
@@ -24,6 +24,8 @@ static size_t const TEMPFILE_MAX_NUM = 16;
 // if we'd create ~10000 "CompatibleDCs", handle handling will fail...
 static int cdc = 999;
 
+// override suppression of repetitive image painting
+bool force_imgs = true;
 
 
 #if CYGWIN_VERSION_API_MINOR >= 74
@@ -240,6 +242,8 @@ winimg_len(imglist *img)
   return img->len ?: img->pixelwidth * img->pixelheight * 4;
 }
 
+#define maxval(type)	((unsigned type)-1 >> 1)
+
 bool
 winimg_new(imglist **ppimg, char * id, unsigned char * pixels, uint len,
            int left, int scrtop, int width, int height,
@@ -274,6 +278,8 @@ winimg_new(imglist **ppimg, char * id, unsigned char * pixels, uint len,
   img->prev = NULL;
   img->strage = NULL;
   img->attr = attr;
+  img->x = maxval(short);
+  img->y = maxval(short);
 
   img->len = len;
   if (len) {  // image format, not sixel
@@ -704,6 +710,14 @@ winimgs_paint(void)
       // do not adjust horizontal scrolling here
 
       int top = img->top - term.virtuallines - term.disptop;
+
+      // suppress repetitive image painting
+      if (left == img->x && top == img->y && !force_imgs)
+        continue;
+      img->x = left;
+      img->y = top;
+      //printf("disp @%d/%d\n", left, top);
+
       if (top + img->height < 0 || top > term.rows) {
         // if the image is scrolled out, serialize it into a temp file
 #ifdef debug_img_list
@@ -912,6 +926,9 @@ winimgs_paint(void)
   }
 
   ReleaseDC(wnd, dc);
+
+  // suppress repetitive image painting
+  force_imgs = false;
 }