+++ wiki/Changelog.md
@@ -3,6 +3,7 @@ Terminal features
   * Fix emoji invisible and blinking attributes.
   * Accept (but ignore) XTMODKEYS subparameters (xterm 398).
   * Fix doubled underline (since 3.7.8).
+  * Avoid repetitive painting of images.
 
 Configuration
   * Changing default setting UnderlineManual=true for consistent line placement.